backend/requirements.txt
# ─── Core Framework ───
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── Database ───
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
aiosqlite==0.20.0

# ─── Redis / Caching ───
redis[hiredis]==5.0.4

# ─── Auth & Security ───
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# ─── AI / ML ───
google-generativeai==0.5.4
numpy==1.26.4

# ─── PDF Generation ───
reportlab==4.2.0

# ─── File Parsing ───
PyMuPDF==1.24.3
python-docx==1.1.2

# ─── Scraping ───
playwright==1.43.0
beautifulsoup4==4.12.3
httpx==0.27.0

# ─── Scheduling ───
apscheduler==3.10.4

# ─── Rate Limiting ───
slowapi==0.1.9

# ─── Validation & Utils ───
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiofiles==23.2.1
